root: general - refactoring to use modern module format
This means using name of module in parent dir, instead of mod.rs in module dir.
The functions shared between the interfaces is now stored in dbus.rs instead
of system_io.rs
