+++ src/platforms.rs
@@ -10,5 +10,7 @@
 //
 // You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
 
-pub mod control_interface;
-pub mod status_interface;
+#![doc = include_str!("platforms/README.md")]
+pub mod platform;
+pub mod universal;
+pub mod universal_components;
