--- src/platforms/mod.rs
@@ -1,16 +1,0 @@
-#![doc = include_str!("./README.md")]
-// This file is part of fpgad, an application to manage FPGA subsystem together with device-tree and kernel modules.
-//
-// Copyright 2025 Canonical Ltd.
-//
-// SPDX-License-Identifier: GPL-3.0-only
-//
-// fpgad is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License version 3, as published by the Free Software Foundation.
-//
-// fpgad is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranties of MERCHANTABILITY, SATISFACTORY QUALITY, or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
-
-pub mod platform;
-pub mod universal;
-pub mod universal_components;
